+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # PHP libs by Murdej
 
-## 1.0 First published version
+## 1.1.0
+
+ - `TreeMaker` link to parent in `TreeItem`
+ - `CollectionDiff` Input collections can be iterable. (Originally only array)
+
+## 1.0.0 First published version
+